refactor: inline pickModel/pickEffort logic and consolidate tests
Inline handleModelCommand and handleEffortCommand bodies directly into
Settings.pickModel() and Settings.pickEffort() — the standalone helpers
are no longer needed. Merge effort.test.ts and model.test.ts into
settings.test.ts, rewriting tests to exercise the Settings class.
