+++ tests/settings.test.ts
@@ -1,15 +1,17 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
-import { Settings, setCachedModels, _resetCachedModels } from "../src/agent/settings.js";
+import { Settings, EFFORT_LEVELS, setCachedModels, _resetCachedModels, findModel } from "../src/agent/settings.js";
 import type { PickResult } from "../src/agent/input.js";
 import { stripAnsi } from "./helpers.js";
 
 const MODELS = [
   { value: "default", displayName: "Default (recommended)", description: "Best for everyday tasks" },
   { value: "opus", displayName: "Opus 4.6", description: "Most capable for complex work" },
+  { value: "haiku", displayName: "Haiku 4.5", description: "Fastest for quick answers" },
 ];
 
 let printed: string[];
 const print = (s: string) => { printed.push(stripAnsi(s)); };
+const noopPick = vi.fn();
 const cancelPick = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
   .mockResolvedValue({ type: "cancelled" });
 
@@ -18,6 +20,8 @@ beforeEach(() => {
   printed = [];
   cancelPick.mockResolvedValue({ type: "cancelled" });
 });
+
+// ── Settings class ───────────────────────────────────────────────────────────
 
 describe("Settings", () => {
   it("initialises model and effort from constructor", () => {
@@ -31,60 +35,271 @@ describe("Settings", () => {
     expect(s.model).toBeUndefined();
     expect(s.effort).toBeUndefined();
   });
+});
 
-  it("pickModel updates model", async () => {
-    setCachedModels(MODELS);
+// ── pickEffort: direct set via argument ──────────────────────────────────────
+
+describe("pickEffort <arg> (direct set)", () => {
+  it("sets effort by known level", async () => {
     const s = new Settings();
-    await s.pickModel("opus", cancelPick, undefined, print);
-    expect(s.model).toBe("opus");
+    await s.pickEffort("low", noopPick, print);
+    expect(s.effort).toBe("low");
+    expect(printed.join("")).toContain("low");
   });
 
-  it("pickModel resets to undefined on 'default'", async () => {
-    const s = new Settings({ model: "opus" });
-    await s.pickModel("default", cancelPick, undefined, print);
-    expect(s.model).toBeUndefined();
+  it("sets each valid level", async () => {
+    for (const level of ["low", "medium", "high", "max"] as const) {
+      const s = new Settings();
+      await s.pickEffort(level, noopPick, print);
+      expect(s.effort).toBe(level);
+    }
   });
 
-  it("pickEffort updates effort", async () => {
-    const s = new Settings();
-    await s.pickEffort("high", cancelPick, print);
-    expect(s.effort).toBe("high");
-  });
-
-  it("pickEffort resets to undefined on 'auto'", async () => {
+  it("'auto' resets to undefined", async () => {
     const s = new Settings({ effort: "low" });
-    await s.pickEffort("auto", cancelPick, print);
+    await s.pickEffort("auto", noopPick, print);
+    expect(s.effort).toBeUndefined();
+    expect(printed.join("")).toContain("auto");
+  });
+
+  it("rejects unknown levels", async () => {
+    const s = new Settings({ effort: "high" });
+    await s.pickEffort("turbo", noopPick, print);
+    expect(s.effort).toBe("high"); // unchanged
+    const output = printed.join("");
+    expect(output).toMatch(/unknown|invalid/i);
+    expect(output).toContain("turbo");
+  });
+});
+
+// ── pickEffort: interactive picker ───────────────────────────────────────────
+
+describe("pickEffort (interactive picker)", () => {
+  it("shows picker with all levels including auto", async () => {
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "cancelled" });
+    const s = new Settings();
+    await s.pickEffort("", pickFn, print);
+    const options = pickFn.mock.calls[0][0] as string[];
+    expect(options.length).toBe(EFFORT_LEVELS.length);
+    expect(options[0]).toMatch(/auto/i);
+  });
+
+  it("selecting auto resets to undefined", async () => {
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "selected", index: 0 }); // auto is first
+    const s = new Settings({ effort: "high" });
+    await s.pickEffort("", pickFn, print);
     expect(s.effort).toBeUndefined();
   });
 
-  it("pickModel with picker updates model", async () => {
-    setCachedModels(MODELS);
+  it("selecting a named level sets the effort", async () => {
+    // EFFORT_LEVELS: auto, low, medium, high, max — "low" is index 1
     const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
-      .mockResolvedValue({ type: "selected", index: 1 }); // opus
+      .mockResolvedValue({ type: "selected", index: 1 });
     const s = new Settings();
-    await s.pickModel("", pickFn, undefined, print);
+    await s.pickEffort("", pickFn, print);
+    expect(s.effort).toBe("low");
+    expect(printed.join("")).toContain("low");
+  });
+
+  it("passes currentIdx matching the active effort", async () => {
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "cancelled" });
+    // "high" is index 3 in [auto, low, medium, high, max]
+    const s = new Settings({ effort: "high" });
+    await s.pickEffort("", pickFn, print);
+    expect(pickFn.mock.calls[0][1]).toBe(3);
+  });
+
+  it("passes currentIdx 0 when no effort set (auto)", async () => {
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "cancelled" });
+    const s = new Settings();
+    await s.pickEffort("", pickFn, print);
+    expect(pickFn.mock.calls[0][1]).toBe(0);
+  });
+
+  it("cancel preserves current effort", async () => {
+    const s = new Settings({ effort: "max" });
+    await s.pickEffort("", cancelPick, print);
+    expect(s.effort).toBe("max");
+  });
+
+  it("selecting already-active auto is a no-op", async () => {
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "selected", index: 0 });
+    const s = new Settings();
+    await s.pickEffort("", pickFn, print);
+    expect(s.effort).toBeUndefined();
+  });
+});
+
+// ── pickModel: direct set via argument ───────────────────────────────────────
+
+describe("pickModel <arg> (direct set)", () => {
+  it("sets model by known alias", async () => {
+    setCachedModels(MODELS);
+    const s = new Settings();
+    await s.pickModel("opus", noopPick, undefined, print);
+    expect(s.model).toBe("opus");
+    expect(printed.join("")).toContain("Opus 4.6");
+  });
+
+  it("warns but accepts unknown model", async () => {
+    setCachedModels(MODELS);
+    const s = new Settings();
+    await s.pickModel("claude-sonnet-4-6-20250514", noopPick, undefined, print);
+    expect(s.model).toBe("claude-sonnet-4-6-20250514");
+    const output = printed.join("");
+    expect(output).toContain("claude-sonnet-4-6-20250514");
+    expect(output).toMatch(/unknown|warning|not recognized/i);
+  });
+
+  it("'default' resets to undefined", async () => {
+    const s = new Settings({ model: "opus" });
+    await s.pickModel("default", noopPick, undefined, print);
+    expect(s.model).toBeUndefined();
+    expect(printed.join("")).toContain("default");
+  });
+
+  it("'sonnet' maps to default", async () => {
+    setCachedModels(MODELS);
+    const s = new Settings({ model: "opus" });
+    await s.pickModel("sonnet", noopPick, undefined, print);
+    expect(s.model).toBeUndefined();
+    expect(printed.join("")).toContain("default");
+  });
+
+  it("accepts value as-is when no cache", async () => {
+    const s = new Settings();
+    await s.pickModel("opus", noopPick, undefined, print);
     expect(s.model).toBe("opus");
   });
 
-  it("pickEffort with picker updates effort", async () => {
-    // EFFORT_LEVELS: auto(0), low(1), medium(2), high(3), max(4)
-    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
-      .mockResolvedValue({ type: "selected", index: 3 }); // high
+  it("fetches models via fetchModelsFn when cache is empty", async () => {
+    const fetchFn = vi.fn().mockResolvedValue(MODELS);
     const s = new Settings();
-    await s.pickEffort("", pickFn, print);
-    expect(s.effort).toBe("high");
+    await s.pickModel("opus", noopPick, fetchFn, print);
+    expect(fetchFn).toHaveBeenCalled();
+    expect(s.model).toBe("opus");
+  });
+});
+
+// ── pickModel: interactive picker ────────────────────────────────────────────
+
+describe("pickModel (interactive picker)", () => {
+  it("shows message when no cache available", async () => {
+    const s = new Settings({ model: "opus" });
+    await s.pickModel("", noopPick, undefined, print);
+    expect(s.model).toBe("opus"); // unchanged
+    expect(printed.join("")).toContain("No model list available");
   });
 
-  it("pickModel cancel preserves model", async () => {
+  it("selecting first entry resets to undefined (default)", async () => {
+    setCachedModels(MODELS);
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "selected", index: 0 });
+    const s = new Settings({ model: "opus" });
+    await s.pickModel("", pickFn, undefined, print);
+    expect(s.model).toBeUndefined();
+  });
+
+  it("selecting a named model sets the model", async () => {
+    setCachedModels(MODELS);
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "selected", index: 1 });
+    const s = new Settings();
+    await s.pickModel("", pickFn, undefined, print);
+    expect(s.model).toBe("opus");
+    expect(printed.join("")).toContain("Opus 4.6");
+  });
+
+  it("passes currentIdx matching the active model", async () => {
+    setCachedModels(MODELS);
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "cancelled" });
+    const s = new Settings({ model: "opus" });
+    await s.pickModel("", pickFn, undefined, print);
+    expect(pickFn.mock.calls[0][1]).toBe(1);
+  });
+
+  it("passes currentIdx 0 when no model set (default)", async () => {
+    setCachedModels(MODELS);
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "cancelled" });
+    const s = new Settings();
+    await s.pickModel("", pickFn, undefined, print);
+    expect(pickFn.mock.calls[0][1]).toBe(0);
+  });
+
+  it("shows model descriptions in options", async () => {
+    setCachedModels(MODELS);
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "cancelled" });
+    const s = new Settings();
+    await s.pickModel("", pickFn, undefined, print);
+    const options = pickFn.mock.calls[0][0] as string[];
+    expect(options[0]).toContain("Best for everyday tasks");
+  });
+
+  it("does not include Other option", async () => {
+    setCachedModels(MODELS);
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "cancelled" });
+    const s = new Settings();
+    await s.pickModel("", pickFn, undefined, print);
+    const options = pickFn.mock.calls[0][0] as string[];
+    expect(options.length).toBe(MODELS.length);
+    expect(options.every(o => !o.includes("Other"))).toBe(true);
+  });
+
+  it("cancel preserves current model", async () => {
     setCachedModels(MODELS);
     const s = new Settings({ model: "opus" });
     await s.pickModel("", cancelPick, undefined, print);
     expect(s.model).toBe("opus");
   });
 
-  it("pickEffort cancel preserves effort", async () => {
-    const s = new Settings({ effort: "max" });
-    await s.pickEffort("", cancelPick, print);
-    expect(s.effort).toBe("max");
+  it("uses displayName from model info, not raw alias", async () => {
+    setCachedModels(MODELS);
+    const pickFn = vi.fn<(options: string[], currentIdx: number) => Promise<PickResult>>()
+      .mockResolvedValue({ type: "cancelled" });
+    const s = new Settings();
+    await s.pickModel("", pickFn, undefined, print);
+    const options = pickFn.mock.calls[0][0] as string[];
+    expect(options[0]).toMatch(/^Default \(recommended\)/);
+    expect(options[1]).toMatch(/^Opus 4\.6/);
+    expect(options[2]).toMatch(/^Haiku 4\.5/);
+  });
+});
+
+// ── findModel matching ────────────────────────────────────────────────────────
+
+describe("findModel", () => {
+  const SDK_MODELS = [
+    { value: "default", displayName: "Default (recommended)", description: "Sonnet 4.6" },
+    { value: "sonnet[1m]", displayName: "Sonnet (1M context)", description: "" },
+    { value: "opus", displayName: "Opus", description: "Opus 4.6" },
+    { value: "opus[1m]", displayName: "Opus (1M context)", description: "" },
+    { value: "haiku", displayName: "Haiku", description: "Haiku 4.5" },
+  ];
+
+  it("exact match on short alias", () => {
+    expect(findModel(SDK_MODELS, "default")?.displayName).toBe("Default (recommended)");
+    expect(findModel(SDK_MODELS, "opus")?.displayName).toBe("Opus");
+    expect(findModel(SDK_MODELS, "haiku")?.displayName).toBe("Haiku");
+    expect(findModel(SDK_MODELS, "opus[1m]")?.displayName).toBe("Opus (1M context)");
+  });
+
+  it("does not match full model IDs via substring", () => {
+    expect(findModel(SDK_MODELS, "claude-sonnet-4-6")).toBeUndefined();
+    expect(findModel(SDK_MODELS, "claude-opus-4-6")).toBeUndefined();
+    expect(findModel(SDK_MODELS, "claude-haiku-4-5")).toBeUndefined();
+  });
+
+  it("returns undefined for no match", () => {
+    expect(findModel(SDK_MODELS, "gpt-4")).toBeUndefined();
+    expect(findModel(SDK_MODELS, "nope")).toBeUndefined();
   });
 });
